backend/requirements.txt
 
fastapi==0.115.7
uvicorn==0.30.6
python-dotenv==1.0.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.5.2
python-multipart==0.0.7
python-decouple==3.8
alembic==1.13.3